+++ app/src/main/java/com/umc/ttoklip/data/repository/town/WriteTogetherRepository.kt
@@ -1,0 +1,14 @@
+package com.umc.ttoklip.data.repository.town
+
+import com.umc.ttoklip.data.model.town.CreateTogethersRequest
+import com.umc.ttoklip.data.model.town.CreateTogethersResponse
+import com.umc.ttoklip.data.model.town.PatchTogetherResponse
+import com.umc.ttoklip.module.NetworkResult
+
+interface WriteTogetherRepository {
+    suspend fun createTogether(body: CreateTogethersRequest): NetworkResult<CreateTogethersResponse>
+    suspend fun patchTogether(
+        postId: Long,
+        body: CreateTogethersRequest
+    ): NetworkResult<PatchTogetherResponse>
+}
