+++ app/src/main/java/com/umc/ttoklip/presentation/hometown/WriteCommunicationViewModel.kt
@@ -1,0 +1,4 @@
+package com.umc.ttoklip.presentation.hometown
+
+interface WriteCommunicationViewModel {
+}
