+++ app/src/main/java/com/umc/ttoklip/presentation/hometown/WriteTogetherViewModel.kt
@@ -1,0 +1,4 @@
+package com.umc.ttoklip.presentation.hometown
+
+interface WriteTogetherViewModel {
+}
